rename deepdti to deepdta
